backend_python/requirements.txt
# Core Flask dependencies
Flask==3.0.0
flask-cors==4.0.0

# Database dependencies
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data processing
pandas==2.1.4
numpy==1.26.2

# Environment variables
python-dotenv==1.0.0

# Wearipedia - Fitbit data extraction
wearipedia

# Google Gemini API for health insights
google-generativeai==0.8.5

# HTTP requests
requests==2.31.0

# Date/time handling
python-dateutil==2.8.2

# Logging
loguru==0.7.2

# Testing dependencies (optional)
pytest==7.4.3
pytest-cov==4.1.0
unittest-xml-reporting==3.2.0

# Development tools (optional)
black==23.12.1
flake8==6.1.0
mypy==1.7.1

# Run Coach dependencies
# Geospatial Analysis
geopandas==0.13.2
shapely==2.0.2
folium==0.14.0
haversine==2.8.1
polyline==2.0.2

# Machine Learning & Scientific Computing
scipy==1.11.4
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0

# Time Series
prophet==1.1.5
statsmodels==0.14.1

# Google Maps integration
googlemaps==4.10.0

# Async HTTP for improved performance
aiohttp==3.9.1

# Optimization
ortools==9.10.4067
pulp==2.7.0

# Visualization dependencies
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1